requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.2
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
idna==2.8
Markdown==3.1
oauthlib==3.0.1
Pillow==6.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.24.1
django-cors-headers==2.4.0
